+++ src/pwshBedrock/Private/Format-MetaTextMessage.ps1
@@ -60,7 +60,11 @@ function Format-MetaTextMessage {
             'meta.llama3-70b-instruct-v1:0',
             'meta.llama3-1-8b-instruct-v1:0',
             'meta.llama3-1-70b-instruct-v1:0',
-            'meta.llama3-1-405b-instruct-v1:0'
+            'meta.llama3-1-405b-instruct-v1:0',
+            'meta.llama3-2-1b-instruct-v1:0',
+            'meta.llama3-2-3b-instruct-v1:0',
+            'meta.llama3-2-11b-instruct-v1:0',
+            'meta.llama3-2-90b-instruct-v1:0'
         )]
         [string]$ModelID,