+++ src/pwshBedrock/Public/Reset-ModelContext.ps1
@@ -66,6 +66,10 @@ function Reset-ModelContext {
             'meta.llama3-1-8b-instruct-v1:0',
             'meta.llama3-1-70b-instruct-v1:0',
             'meta.llama3-1-405b-instruct-v1:0',
+            'meta.llama3-2-1b-instruct-v1:0',
+            'meta.llama3-2-3b-instruct-v1:0',
+            'meta.llama3-2-11b-instruct-v1:0',
+            'meta.llama3-2-90b-instruct-v1:0',
             'mistral.mistral-7b-instruct-v0:2',
             'mistral.mistral-large-2402-v1:0',
             'mistral.mistral-large-2407-v1:0',
@@ -146,6 +150,11 @@ function Reset-ModelContext {
                             $model -eq 'meta.llama3-70b-instruct-v1:0' -or
                             $model -eq 'meta.llama3-1-8b-instruct-v1:0' -or
                             $model -eq 'meta.llama3-1-70b-instruct-v1:0' -or
+                            $model -eq 'meta.llama3-1-405b-instruct-v1:0' -or
+                            $model -eq 'meta.llama3-2-1b-instruct-v1:0' -or
+                            $model -eq 'meta.llama3-2-3b-instruct-v1:0' -or
+                            $model -eq 'meta.llama3-2-11b-instruct-v1:0' -or
+                            $model -eq 'meta.llama3-2-90b-instruct-v1:0' -or
                             $model -eq 'mistral.mistral-7b-instruct-v0:2' -or
                             $model -eq 'mistral.mixtral-8x7b-instruct-v0:1' -or
                             $model -eq 'mistral.mistral-large-2402-v1:0' -or