+++ docs/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.2.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.22.0] - **BREAKING CHANGES**
+## [0.26.0] - **BREAKING CHANGES**
 
 - Module changes:
     - stability.ai
@@ -22,7 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Cohere
         - `Invoke-CohereCommandRModel` - minor corrections to debug output
     - Meta
-        - `Invoke-MetaModel` - minor corrections to debug output and help
+        - `Invoke-MetaModel`
+            - Added support for new 3.2 models: `meta.llama3-2-1b-instruct-v1:0`, `meta.llama3-2-3b-instruct-v1:0`, `meta.llama3-2-11b-instruct-v1:0`,  `meta.llama3-2-90b-instruct-v1:0`
+            - minor corrections to debug output and help
     - Mistral
         - `Invoke-MistralAIModel` - minor corrections to debug output and help. Adjusted Max token limit validation.
 - Build changes: